Fix an issue with modals and z-index problems
In this case, the modal was attached to the sticky header of the page.
The sticky header had a z-index of 1, which overrode the z-index of the
modal, allowing the modal backdrop to appear above everything.

The only solution was to make the assessment-selection controller cover
the entire page so that the modal dialogue could be made a top-level
element.