Iterate over the benchmarks when displaying the draft plan
We have been iterating over the activity map. Aside from an undetected
capacity duplication error, this also meant that we would never even
display a section for a benchmark that never existed in the activity
map.

Iterating over the entire draft plan requires that the code handles
places where we query the activity map for benchmarks that are absent.