update hamlit to support prettier's haml
prettier checks the `haml` version, but doesn't check the `hamlit`
version. we have a haml version new enough, and it seems to be getting
used in development, but in staging/production it seems that we are
using an older version of `hamlit`, that does not support multiline
attributes, and that's why we're getting the errors.