build action list first, skip pages for no actions
we cannot, it turns out, use the index to check whether a worksheet
exists yet, because we don't use the index to insert the worksheet.
if you just plop new sheets on to the end, but skip sheets if there
are no actions, you might not have sheets that line up exactly with
indexes. and then excel will freak out, because you have added pages
to sheets that you never created. so we will stop doing that.