WIP: break up terraform from one file for everything to three files: one for centralized config, one for the main infrastructure, and one for devops-related stuff. next will make use of terraform workspaces to have environments.
WIP on Private Link for ACR (supposedly should be working now) and DB (not yet up)..

Got ACR working with docker authentication but publicly accessible so will have to continue to keep secrets out of docker images. add Postgres DB server instance on a Private Endpoint, is working via VPN only which is good. VPN via MacOS is working and stable so thats huge, works via IP but does not seem to be able to resolve internal domains/FQDNs. Need to use a General Purpose sku for the Postgres DB server to get network/IP/private-endpoints. currently using Private DNS Zone for DB but does not seem to make a difference with VPN/remote conns.

Change branch for the build pipeline to use my current topic branch, azure--virtual-network-and-subnets--179242298

Now with VNet and Subnets in place, and Private Link for the DB, bring all the other cloud resources back up. App Service and Build Pipeline are there but still not fully working as before, a known issue which we'll tackle in a separate story.