Hardcode the resource groups in azure-pipelines.yml
These can't change as easily now. There's no going back and
forth between them. I used to allow them to be modified because
it was plausible that we could deploy to one or the other.
Spreading everything across the two of them means there's no
benefit anymore to leaving them parameterized.